Bcc 22 marketplace (#10)
* bcc-22 update database scheme and database creation

* bcc-22 implemented player, inv and cards

* bcc-22 offer functionality and exceptions

* Bypass Auth

+++ beancards/src/main/java/com/bbdgrads/beancards/repositories/TradeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bbdgrads.beancards.entities.Trade;
 
-public interface TradeRepository extends JpaRepository<Trade, Long> {
+public interface TradeRepository extends JpaRepository<Trade, Integer> {
 
 }
